+++ R/namespace-env.r
@@ -1,6 +1,6 @@
 #' Return the namespace environment for a package.
 #'
-#' Contains all (exported and non-exported) objects, and is a descendent of
+#' Contains all (exported and non-exported) objects, and is a descendant of
 #' `R_GlobalEnv`. The hierarchy is `<namespace:pkg>`,
 #' `<imports:pkg>`, `<namespace:base>`, and then
 #' `R_GlobalEnv`.
@@ -21,7 +21,10 @@ ns_env <- function(package) {
 }
 
 ns_path <- function(package) {
-  getNamespaceInfo(asNamespace(package), "path")
+  ns <- asNamespace(package)
+  if (isBaseNamespace(ns))
+    return(path.package(package))
+  getNamespaceInfo(ns, "path")
 }
 
 # Create the namespace environment for a package