+++ NEWS.md
@@ -1,4 +1,11 @@
-# pkgload 1.0.1.9000
+# pkgload (development version)
+
+# pkgload 1.0.2
+
+* `shim_question()` now works for topics from the R base package that are passed with the double colon operator (e.g. `base::min`) (@mdequeljoe, #99).
+
+* `load_all()` now allows using explicitly qualified, exported names in test
+  helpers (@klmr, #95).
 
 * `load_all()` gains a `compile` argument which controls more finely whether to
   compile the code or not. The `recompile` argument is now deprecated and will